Rename core-lib tests to fairmq-lib tests
